+++ include/widget_switch.hpp
@@ -6,7 +6,7 @@
 
 class WidgetSwitch : public Widget {
 public:
-  WidgetSwitch(LuaSaveParser *saveParser, u64 onValue, u64 offValue);
+  WidgetSwitch(LuaSaveParser *saveParser, s32 onValue, s32 offValue);
   WidgetSwitch(LuaSaveParser *saveParser, std::string onValue, std::string offValue);
 
   ~WidgetSwitch();
@@ -17,7 +17,7 @@ public:
   void onTouch(touchPosition &touch);
 
 private:
-  u64 m_intOnValue, m_intOffValue;
+  s32 m_intOnValue, m_intOffValue;
   std::string m_strOnValue, m_strOffValue;
 
 };
